NotifymeV1/apps/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
langchain==0.0.340
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
psycopg2-binary==2.9.9
redis==5.0.1
